rauc-native: move native deploy task to extra recipe
This fixes sstate issues when trying to accelerate do_deploy task from
sstate for non-native packages. These did not install anything to deploy
dir which made sstate_installpkg fail with:

|      0317:    for state in ss['dirs']:
|      0318:        prepdir(state[1])
|  *** 0319:        os.rename(sstateinst + state[0], state[1])
|      0320:    sstate_install(ss, d)
|      0321:
|      0322:    for plain in ss['plaindirs']:
|      0323:        workdir = d.getVar('WORKDIR', True)
| Exception: OSError: [Errno 2] No such file or directory

By adding the deploy task only for the native recipe (where it is used) this
issue is fixed.

This also changes install location from ${DEPLOY_DIR_TOOLS} to ${DEPLOYDIR}
which is the one used as sstate-inputdirs.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>