rauc: explicitly set systemd and dbus paths for configure
If not set, these pathes will be determined by running pkg_config. As
systemd is not in the DEPENDS of rauc, it could happen that rauc was
configured before systemd was built, leading to an empty value for the
unitdir path and thus the rauc.service file was not built. This led to

  ERROR: Function failed: SYSTEMD_SERVICE_rauc-service value rauc.service
does not exist

Now, the pathes are set explicitly which makes the package build
correctly but avoids a build dependency on systemd.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>