classes/bundle: fix error in sstate handling
When changing a bundle recipe in a way that would possibly allow bitbake to use
cached data from the sstate, it fails during do_packagedata_setscene with the
following stacktrace:

| File: '[...]/meta/classes/sstate.bbclass', lineno: 320, function: sstate_installpkg
|      0316:
|      0317:    for state in ss['dirs']:
|      0318:        prepdir(state[1])
|      0319:        bb.warn("Rename form {} to {}".format(sstateinst + state[0], state[1]))
|  *** 0320:        os.rename(sstateinst + state[0], state[1])
|      0321:    sstate_install(ss, d)
|      0322:
|      0323:    for plain in ss['plaindirs']:
|      0324:        workdir = d.getVar('WORKDIR', True)
| Exception: OSError: [Errno 2] No such file or directory

Bitbake tries to unpack and move non-existing data from the packagedata task.
As no package will be created from a bundle recipe (note that do_package is set
to 'noexec') the sstate is empty and thus unpacking and moving fails.

This will be fixed by simply not executing the useless task
`packagedata`. As `package_qa` is not required, too, it can also be
deactivated.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>