classes/genimage: place genimage task after do_configure to fix deps
When having a genimage-based image recipe, this might require some extra
host tools that might not be part of the system but have to be build
from the BSP.

In order to achive this, you would have to add, e.g.

  DEPENDS += "genext2fs-native"

to your image recipe.

As this will make image-recipe:do_configure depend on
genext2fs-native:do_populate_sysroot while the `genimage` task was
placed only after `image-recipe:do_unpack`, there was no guarantee that
the tools required for the genimage task were build before the
do_genimage task ran.
In many cases, this will make the build fail because of missing host
tools.

To fix this, the do_genimage task must be placed *after* do_configure, to
ensure that the dependency handling places the tools do_populate_sysroot
task *before* the images do_genimage task.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>