barebox: install bindir only if target tools enabled
Prior to this change, the directory ${D}${bindir} was created
unconditionally and not added to any package. This led to packaging
error due to QA check [installed-vs-shipped].

This is fixed now by creating the install dir conditionally for each
target tool. Multiple calls of install on the same dir will not cause
any issues.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>