barebox: add bison and flex build dependencies
Both lex/yacc or compatibles are now dependencies of the barebox build
process, same as the kernel's. They are used since 273dbe5f8 contained
in barebox v2019.02.0 for generating the Kconfig lexer and parser.

Adding flex-native and bison-native to DEPENDS addresses this and fixes
the currently failing barebox v2019.02.0 do_configure stage:

	| DEBUG: Executing shell function do_configure
	| NOTE: make CROSS_COMPILE=arm-ptx-linux-gnueabi- oldconfig
	|   YACC    scripts/kconfig/zconf.tab.c
	| /bin/sh: 1: bison: not found
	| scripts/Makefile.lib:194: recipe for target 'scripts/kconfig/zconf.tab.c' failed
	| make[1]: *** [scripts/kconfig/zconf.tab.c] Error 127
	| Makefile:416: recipe for target 'oldconfig' failed
	| make: *** [oldconfig] Error 2

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>