classes: genimage: use [dirs] varflag to setup build dir
The bitbake task varflag dirs allows to specify directories to create
before the task runs. The last directory will be used as the workdir for
the current task.

This changes the tasks workdir from ${WORKDIR} to ${B} which is actually
the same for now.
It also allows to remove the manual 'cd' from the beginning of the
do_genimage task that potentially caused issues for users that attempt
to prepend it and landed in the wrong directory.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>