add ptx-profile package only when 'ptx' distro is active
The ptx layer should not unconditionally manipulate any existing package
to be compatible with the default YP guidelines.

To have ptx-profile only active when 'ptx' distro is selected, we add it
via DISTRO_EXTRA_RDEPENDS. So it will make its way into images that use
packagegroup-base, which is recommended, anyway.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>