barebox: prevent BAREBOX_IMAGES check from failing unintentionally
The intention of the statement was to trigger a bberror if
no image could be found and ${BAREBOX_IMAGES} contains anything else
than the default glob '*.img'.
However, when the first expression evaluates to false and thus not
triggers the bberror, it will in turn let the script silently fail with
exit 1 which is not the intention here.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>