barebox: populate new BUILDSYSTEM_VERSION option
Since v2020.11.0 barebox now supports a make BUILDSYSTEM_VERSION= option
that allows embedding a string into the barebox binary that is printed
to console and readable in barebox and Linux via imd.

This is useful because the barebox release may not reflect auxiliary
information like the configuration and the built-in environment.

Populate this variable with the git commit-ish of the git repository
containing the barebox recipe. If there's none, an empty string is
passed, which is the default behavior.

This can be enabled unconditionally for older barebox versions as well,
because unknown make variables are ignored.

Recipes that want to customize the version, can override
BAREBOX_BUILDSYSTEM_VERSION to their preferred value.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>