genimage: add deploy task after genimage
Instead of coding the deployment into the genimage task, use genimage as
a separate task which will only create the requested image, instead of
directly copying it into the ${DEPLOYDIR}. This is necessary since
fakeroot tasks add to the SSTATE, will try to record the uid and gid of
the files, however since PSEUDO_IGNORE_PATH was introduced into oe-core,
files within the deploy directory are ignored. This leads to an error
when trying to add files from the ${DEPLOYDIR} into the SSTATE:

  WARNING: simple-sd-image-1.0-r0 do_genimage: KeyError in ./deploy-simple-sd-image
  File: '/home/phoenix/build/YOCTO.BSP-Pengutronix-OP-TEE/meta/lib/oe/sstatesig.py', lineno: 592, function: OEOuthashBasic
       0588:
       0589:                update_hash("\n")
       0590:
       0591:            # Process this directory and all its child files
   *** 0592:            process(root)
       0593:            for f in files:
       0594:                if f == 'fixmepath':
       0595:                    continue
       0596:                process(os.path.join(root, f))
  File: '/home/phoenix/build/YOCTO.BSP-Pengutronix-OP-TEE/meta/lib/oe/sstatesig.py', lineno: 555, function: process
       0551:                    add_perm(stat.S_IXOTH, 'x')
       0552:
       0553:                if include_owners:
       0554:                    try:
   *** 0555:                        update_hash(" %10s" % pwd.getpwuid(s.st_uid).pw_name)
       0556:                        update_hash(" %10s" % grp.getgrgid(s.st_gid).gr_name)
       0557:                    except KeyError:
       0558:                        bb.warn("KeyError in %s" % path)
       0559:                        raise
  Exception: KeyError: 'getpwuid(): uid not found: 1000'

We copy every every generated image to the deploy directory to not break
the existing behaviour, where genimage generated every image file in the
deploy directory directly.

Fixes https://github.com/pengutronix/meta-ptx/issues/60

Signed-off-by: Rouven Czerwinski <rouven@czerwinskis.de>