barebox: fix building userspace programs
Barebox' commit [1] contained in v2020.08.0 for the first time broke building
the userspace programs for the target system (bareboxenv & Co.). Fix that by
offering the needed compiler flags not only in the present parameter
"TARGETCFLAGS", but additionally in the newly introduced one "userccflags".

Note that this might not be a definitive solution as it's unclear whether
userccflags is meant as user API in Linux (which barebox Kbuild is synchronized
with).

The same issue should happen with kernel samples, which are also target tools.
Those seem not to be handled by poky. Selftests are supported, but those have
their own Makefile and don't result as part of the normal Kbuild procedure.

[1] c008d836d8f4 ("scripts: use 'userprogs' to build programs for target")

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>