confuse: add patch to suppress path search for absolute paths
Naively adding an "--includepath ${WORKDIR}" to the genimage command
line breaks the build, because the main configuration file is given
via the absolute path ${B}/.config.tmp, and when there's a search path
set, libconfuse only looks for files by path-joining each of the
search directories with the given path. But that's the wrong thing to
do for absolute paths.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>