genimage.bbclass: allow recipe to pass extra options to genimage
It's useful to be able to pass extra custom options to genimage,
e.g. when one is building a newer genimage than the one defined in
meta-ptx to gain access to new features. For example, I wish to make
use of both --includepath (which is currently in genimage master) and
the just-proposed --configdump option.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>