genimage.bbclass: add --includepath option
It is currently somewhat awkward to use include() from within
genimage.config, since one has to arrange for all the fragments to be
copied to ${B}; it is not enough to simply have them listed in
SRC_URI. So currently, we have this in our image recipe:

do_genimage:prepend() {
    cp ${S}/*.config ${B}
}
do_genimage:append() {
    # do_install takes everything in $B and puts it in $DEPLOYDIR. The
    # .config fragments serve no purpose there.
    rm -f ${B}/*.config
}

While the first is somewhat tolerable, the need for the second
fragment is rather annoying.

Now that genimage (since v15, and built against new enough libconfuse
which we do) has an --includepath option, add that to the command line
and point it at ${S} aka ${WORKDIR}, making include() work
out-of-the-box.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>
Signed-off-by: Enrico Jorns <ejo@pengutronix.de>