genimage.bbclass: set LICENSE with '?=' to avoid forcing MIT license.
While genimage itself is MIT-licensed, images created with genimage using
the 'inherit genimage' should not be forced to be MIT-licensed. The
standard for bbclass files which set LICENSE is to set it with '?=' so
the recipe can override it. Otherwise (with '=') it isn't possible for
the recipe to override it.

Signed-off-by: David Norris <david.norris@blueberrydaq.com>
(cherry picked from commit 87c548cc48a17dbbe3e934fc1d85823255bfeec3)
Signed-off-by: Enrico Jörns <ejo@pengutronix.de>