classes: move classes to match bitbake's class scope functionality
Starting with bitbake-2.2 which is part of yocto-4.1 ("langdale") classes can
explicitly be classified to be designed for global usage in the overall
configuration space or recipe-wise usage by moving them into the directories
"classes-global" or "classes-recipe" respectively (see Bitbake's commit [1] /
Poky's commit [2]).

Let's do this for meta-ptx' existing classes.

[1] f33ce7e742f4 ("BBHandler/cooker: Implement recipe and global classes")
[2] 7bd328f9d24b ("bitbake: BBHandler/cooker: Implement recipe and global classes")

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>