classes-recipe: add platsch.bbclass
Generalize needed functionality to automatically convert images to be displayed
by platsch into a dedicated class. Currently only the RGB565 format is supported.

A recipe consuming that class could look as simple as e.g.

    SRC_URI = " \
        file://splash-1280x800.png \
        file://splash-800x480.png \
    "

    inherit platsch

    SPLASH_IMAGES = " \
        splash-1280x800.png \
        splash-800x480.png \
    "
    SPLASH_FORMAT = "RGB565"

and would produce a package containing the raw bitmaps in RGB565 format in the
default location ${datadir}/platsch (which normally is /usr/share/platsch) as

    /usr
    └── share
        └── platsch
            ├── splash-1280x800-RGB565.bin
            └── splash-800x480-RGB565.bin

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>