genimage: enable upstream check
This sets UPSTREAM_CHECK_REGEX properly.
Despite the 'github-releases' classes defaults dot not really meet the
requirements of genimage, use it to document it uses releases from
GitHub and potentially benefit from further extensions.

The default UPSTREAM_CHECK_REGEX does not work for us since it expects
at least a minor version delimited by a dot.

The regex used here is exactly the same as the default, with just the
'+' after '(\.\d+)' replaced by a '*' to make the minor version
optional.

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>