genimage: remove recipe
Since c3a11589 ("genimage: add new recipe "), the genimage recipe is
part of meta-openembedded's meta-oe and can be removed from meta-ptx.

Same applies to the 'confuse' recipe which was there for genimage only.
The 'libconfuse' pendant was moved from meta-networking to meta-oe and
prepared for being compatible with genimage.

Since genimage.bbclass is still part of meta-ptx, add meta-oe
('openembedded-layer') to the list of layer dependencies.

Signed-off-by: Enrico Jörns <ejo@pengutronix.de>