.github/workflows/schedule-builds: Run scheduled jobs via workflow_dispatch
We currently only test the master branch on a regular schedule and not the
different release branches (like e.g. scarthgap or styhead).

Also testing these other branches is complicated by the following detail
in the GitHub action documentation[1]:

> Scheduled workflows will only run on the default branch

This means we need some sort of workaround to run scheduled jobs on
different branches. This commit adds a scheduled job on the master branch
to trigger jobs on other branches via the `workflow_dispatch` event.

Build the master branch as well as all current release branches.

[1]: https://docs.github.com/en/actions/writing-workflows/choosing-when-your-workflow-runs/events-that-trigger-workflows#schedule

Signed-off-by: Leonard Göhrs <l.goehrs@pengutronix.de>